requirements.txt
fastapi
uvicorn
google-generativeai
manim
google-genai
python-multipart
